requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
python-binance
dash
pandas
plotly
